refactor(frontend): extract shared BboxCrop; add group thumbnail types
