+++ frontend/src/pages/SequenceGroupAnnotatePage.tsx
@@ -13,6 +13,7 @@ import {
   X,
 } from 'lucide-react';
 import { apiClient } from '@/services/api';
+import { Bbox, BboxCrop } from '@/components/annotation/BboxCrop';
 import { useDetectionImage } from '@/hooks/useDetectionImage';
 import { usePersistedTabState } from '@/hooks/usePersistedTabState';
 import { useState } from 'react';
@@ -43,36 +44,8 @@ function memberIsAnnotated(m: SequenceGroupMember): boolean {
   );
 }
 
-type Bbox = [number, number, number, number];
-
 function isValidBox([x1, y1, x2, y2]: Bbox): boolean {
   return x2 > x1 && y2 > y1;
-}
-
-// Zoomed view of a single image centered on `box`. The same (already cached)
-// detection image is reused and magnified with a CSS transform — no second
-// fetch, no canvas — so small objects are legible next to the full frame.
-function BboxCrop({ url, box }: { url: string; box: Bbox }) {
-  const [x1, y1, x2, y2] = box;
-  const cx = (x1 + x2) / 2;
-  const cy = (y1 + y2) / 2;
-  // Show the box plus a margin of one box-size on each side (region ≈ 3×),
-  // then zoom so the whole region fits the cell; cap at 8× for tiny boxes.
-  const regionW = Math.min(1, Math.max(x2 - x1, 0.001) * 3);
-  const regionH = Math.min(1, Math.max(y2 - y1, 0.001) * 3);
-  const zoom = Math.min(1 / regionW, 1 / regionH, 8);
-
-  return (
-    <img
-      src={url}
-      alt=""
-      className="absolute inset-0 w-full h-full object-cover"
-      style={{
-        transformOrigin: `${cx * 100}% ${cy * 100}%`,
-        transform: `translate(${(0.5 - cx) * 100}%, ${(0.5 - cy) * 100}%) scale(${zoom})`,
-      }}
-    />
-  );
 }
 
 function MemberCard({
