+++ frontend/src/types/api.ts
@@ -288,6 +288,14 @@ export interface SequenceGroupMember {
   first_detection_algo_predictions: AlgoPredictions | null;
 }
 
+export interface SequenceGroupThumbnail {
+  detection_id: number;
+  url: string;
+  // Crop box for the thumbnail; null when the frame has no valid
+  // prediction boxes — fall back to the group's representative_bbox.
+  bbox_xyxyn: [number, number, number, number] | null;
+}
+
 export interface SequenceGroupListItem {
   id: number;
   camera_id: number;
@@ -305,6 +313,8 @@ export interface SequenceGroupListItem {
   labeled_at: string | null;
   created_at: string;
   member_count: number;
+  // Up to 3 member previews (first/middle/last member by recorded_at).
+  thumbnails: SequenceGroupThumbnail[];
 }
 
 export interface SequenceGroupStats {
